+++ src/main/java/umc/cockple/demo/domain/chat/presentation/rest/api/ChatApiTag.java
@@ -1,0 +1,14 @@
+package umc.cockple.demo.domain.chat.presentation.rest.api;
+
+import io.swagger.v3.oas.annotations.tags.Tag;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@Tag(name = "Chat", description = "채팅 관리 API")
+public @interface ChatApiTag {
+}
